Added Printer Options

+++ IspDataDb/dbo/Tables/CodeGroup.sql
@@ -1,5 +1,5 @@
 ﻿CREATE TABLE [dbo].[CodeGroup] (
-    [IdNo]             TINYINT        IDENTITY (1, 1) NOT NULL,
+    [IdNo]             INT            IDENTITY (1, 1) NOT NULL,
     [CodeGroupCode]    VARCHAR (4)    NULL,
     [CodeGroupName]    NVARCHAR (50)  NULL,
     [CodeGroupNameAra] NVARCHAR (50)  NULL,
@@ -7,6 +7,8 @@
     [DateTimeStamp]    ROWVERSION     NULL,
     CONSTRAINT [PK_CodeGroup] PRIMARY KEY CLUSTERED ([IdNo] ASC)
 );
+
+
